Rename arrow icon files
