+++ src/components/icons/arrow-right-icon.tsx
@@ -1,5 +1,5 @@
 import { ComponentProps } from 'react';
-import ArrowLeftIcon from './ArrowLeftIcon';
+import ArrowLeftIcon from './arrow-left-icon';
 
 export default function ArrowRightIcon(props: ComponentProps<'svg'>) {
   return <ArrowLeftIcon {...props} style={{ ...props.style, transform: 'rotate(180deg)' }} />;
